requirements.txt
pandas>=1.5.2
numpy>=1.23.5
scipy>=1.10.0
matplotlib>=3.6.2
statsmodels>=0.13.5
parameterized>=0.8.1
qablet_contracts>=0.2.2
qablet-basic>=0.3.2
pyarrow>=13.0.0